custom-elements/registering-custom-elements/unresolved-element-pseudoclass/unresolved-element-pseudoclass-test-003.html

<!DOCTYPE html>
<html>
<head>
<title>The :unresolved pseudoclass matching with Document.querySelectorAll()</title>
<meta name="author" title="Aleksei Yu. Semenov" href="mailto:a.semenov@unipro.ru">
<meta name="assert" content="The :unresolved pseudoclass must match all custom elements whose created callback has not yet been invoked">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#unresolved-element-pseudoclass">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">
test(function() {
    var doc = newHTMLDocument();
    doc.body.innerHTML = '<x-a-1 id="xa"></x-a-1>';
    var queryResult = doc.querySelectorAll(':unresolved');

    assert_equals(queryResult.length, 1,
        'Unregistered custom element should be accessible by :unresolved');

    assert_equals(queryResult.item(0).localName, 'x-a-1', 'Wrong custom element');

}, 'Document.querySelectorAll(): Single unresolved custom element should be accessible by :unresolved');


test(function() {
    var doc = newHTMLDocument();
    doc.body.innerHTML = '<x-a-2 id="xa"></x-a-2>';
    var queryResult = doc.querySelectorAll(':unresolved');

    assert_equals(queryResult.length, 1,
        'Unresolved custom element should be accessible by :unresolved');

    assert_equals(queryResult.item(0).localName, 'x-a-2', 'Wrong custom element');

    doc.registerElement('x-a-2');
    queryResult = doc.querySelectorAll(':unresolved');

    assert_equals(queryResult.length, 0,
        'Resolved custom element should not be accessible by :unresolved');

}, 'Document.querySelectorAll(): Single resolved custom element should not be accessible by :unresolved');


test(function() {
    var doc = newHTMLDocument();
    doc.body.innerHTML = '<x-a-3 id="xa3"></x-a-3>'
        + '<x-a-4 id="xa4"></x-a-4>';
    var queryResult = doc.querySelectorAll(':unresolved');

    assert_equals(queryResult.length, 2,
        'Unregistered custom element should be accessible by :unresolved');

    assert_equals(queryResult.item(0).localName, 'x-a-3', 'Wrong first custom element');
    assert_equals(queryResult.item(1).localName, 'x-a-4', 'Wrong second custom element');

    doc.registerElement('x-a-3');
    queryResult = doc.querySelectorAll(':unresolved');

    assert_equals(queryResult.length, 1,
        'Unresolved custom element should be accessible by :unresolved');

    assert_equals(queryResult.item(0).localName, 'x-a-4', 'Wrong custom element');

}, 'Document.querySelectorAll(): Several unresolved custom elements should be accessible by :unresolved. ');


test(function() {
    var doc = newHTMLDocument();

    var customElement = doc.createElement('x-a-5');
    doc.body.appendChild(customElement);

    var queryResult = doc.querySelectorAll(':unresolved');

    assert_equals(queryResult.length, 1,
        'Unregistered custom element should be accessible by :unresolved');

    assert_equals(queryResult.item(0).localName, 'x-a-5', 'Wrong custom element');

}, 'Document.querySelectorAll(): Unresolved custom element, created via Document.createElement(), '+
    'should be accessible by :unresolved');


testInIFrame('../../resources/x-element.html', function(doc) {

    var queryResult = doc.querySelectorAll(':unresolved');

    assert_equals(queryResult.length, 1,
        'Unregistered custom element should be accessible by :unresolved');

    assert_equals(queryResult.item(0).localName, 'x-element', 'Wrong custom element');
}, 'Document.querySelectorAll(): Unresolved custom element should be accessible by :unresolved '+
    'in loaded document');
</script>
</body>
</html>
